Improve management of relevant terms in TheoryStrings (#9149)
This makes the computation of relevant terms cached per full effort check.
It also corrects an issue where terms that are not relevant for model building were not considered relevant.

Finally, it makes it so that string reductions are only applied for active and relevant terms. This change currently has very little impact since we eagerly add lemmas for all preregistered terms (i.e. preregistered terms and terms in assertions are roughly equivalent).

I did an SMT-LIB run with this change and there was no difference in solved benchmarks or significant differences in time.

This is in preparation for further testing for making our preregistration lemmas done lazier.