Fixes for theory combination + non-linear arithmetic (#9175)
This makes arithmetic substitution for arithmetic getEqualityStatus (e.g. a substitution that does not traverse symbols that do not belong to the theory). This ensures that getEqualityStatus does not look inside terms that don't belong to arithmetic, which can have a subtle impact on theory combination.

It also makes it so that we never assert values for non-linear multiplication terms to the model. These values correspond to the model for their linear abstraction, which is not guaranteed to be accurate. The nonlinear extension is not properly repairing some of these in the arithmetic model, but regardless asserting values for these terms simply should be avoided in collectModelInfo.

Fixes #9164.