Refactor lazy term registration options in strings (#9187)
This refactors calls to TermRegistry::registerTerm in strings. The option --no-strings-eager-reg can now be used to do term registration slightly lazier (when terms appear in asserted literals). This replaces and simplifies a few deprecated variants of this feature.

The behavior is not changed in this PR. An SMT-LIB run on strings shows mixed performance with this option (+68-81) on QF_S, QF_SLIA, QF_SNIA with a 60 second timeout.