Do not assert arithmetic model when non-linear is not sat (#9206)
This avoids constructing inconsistent (intermediate) models when using the non-linear arithmetic solver.

The non-linear extension of arithmetic often "buffers" lemmas when it finds the current assertions are not satisfiable. For the sake of fairness, it does not send them until LAST_CALL effort (after models have been constructed), so that other theories have an opportunity to add lemmas as well. The problem is the model may be inconsistent in such cases.

This PR changes TheoryArith::collectModelInfo for the case when such lemmas are buffered. In particular, we only assert the equality engine (the default implementation).

Fixes #9183. The second benchmark on that issue times out.