Add option for incremental evaluation tests for quantifier instantiation (ieval) (#9158)
This adds an option ieval for incremental instantiation evaluation tests. These tests are run as part of the implementation InstMatch objects (currently being used by E-matching). When the ieval option is enabled, we consult the appropriate instantiation evaluator to check if the match is feasible.

It furthermore extends conflict-based instantiation to maintain InstMatch objects, hence allowing entailment to be tested eagerly and incrementally in conflict-based instantiation.

I am keeping this disabled by default still for now, but it will likely be enabled by default soon after further testing.

--ieval=use is +209-132 on SMT-LIB quantified logics compared with default. --ieval=use-learn is slightly worse than --ieval=use (-14+8).