[proof] [prop] Cache prop proof per check-sat (#9242)
Avoids potential issues when requesting to build the prop engine proof twice in the same check-sat call. This is done by caching the computed proof nodes. Note it's two proof nodes since a prop engine proof may be requested connecting the SAT proof or not to the proofs of its leaves stored in the ProofCnfStream.

Fixes #9156 and its duplicate #9172.