Use separate subsolver for synthesizing rewrite rules from input (#9237)
This experimental option causes cvc5 to synthesize rewrite rules based on the input. The code for this is a preprocessing pass which used to rewrite the input to a sygus conjecture. This is confusing since it changes the semantics of the input. We now call a separate subsolver.

This also cleans up set defaults. The options required for synthesizing rewrite rules are now applied locally to the subsolver spawned by the preprocessing pass.

We now terminate with an exception in the rare case when rewrite rule synthesis terminates.

This is required for further enhancements to the sygus solver for answering infeasible.