Simplify assertions interface (#9239)
Removes some interfaces that are used by miplib and ITE simplification passes, which are experimental. "real assertions end" is now interpreted to be the actual last index of the assertions, instead of tracking this manually. It seems this was an optimization for not traversing the same assertions twice when using repeated simplification. This optimization should be managed locally by these passes (if necessary). I will do a run on QF_LIA to test this.

This is work towards making our assertions more flexible, which will allow deep restarting to be potentially compatible with models and incremental.