Not allowing bound variables under uninterpreted functions with bv-to-int (#9247)
Solves #9071 .

The current implementation of int-blasting does not support bound variables under uninterpreted functions. The reason is that these function applications must be guarded to be in the right range, but this range-constraint should be under the appropriate quantifier, and not in the top level formula.

This PR throws an error when this case is reached.