Towards decoupling Assertions from AssertionPipeline (#9251)
This is work towards decoupling Assertions (the current set of user assertions) from AssertionPipeline (the set of assertions we are currently preprocessing).

This makes the preprocessor only depend on the latter. It changes the output of -o post-asserts slightly by making it more complete: we now print all top-level substitutions from Env instead of all define-fun from the assertions list.