Add theory interface for marking unsoundness (#9238)
We have a mechanism for marking "incomplete", e.g. replace "sat" with "unknown" ("incomplete" is changed to "model unsound" in this PR).

This adds the dual feature for marking "refutation unsound", e.g. replace "unsat" with "unknown".

This is useful if a theory decides to send a lemma that is unsound.

We have one such use case for unsound lemmas currently: the SyGuS extension of quantifiers when it fails to show an enumerated candidate is definitely not a solution but fails to find a counterexample (this may happen in undecidable background theories), it will simply block it to make progress. Currently we are marking "incomplete", but in fact this should mark "unsound". As such, the current SyGuS solver may incorrectly answer "infeasible" instead of "fail", due to the recent changes for interpreting "unsat" as "infeasible". This is fixed in this PR, a regression is added.