[proofs] [alethe] Alethe let binding
Adds a specific let binder for Alethe proofs. It is necessary because sharing is
represented via the ":named" attribute of SMT-LIB, where the first
occurrence (in a post-order visit) is wrapped in a attribute to declare a
name (i.e., a fresh variable) and the next occurrences are replaced by that
variable.