Do not report infeasible for smart enumeration + any-constant constructors (#9262)
We currently (intentionally) overprune any-constant constructor model values, by not considering the concrete model value of their arguments when using smart enumeration in SyGuS solver. This incompleteness was not accounted for yet, hence we could return infeasible incorrectly for grammars that had finitely many values modulo any-constant argument values. cvc5/cvc5-projects#222 was an example.

This also corrects a bug where the SyGuS conjecture itself was being included in the assertions to check with --check-synth-sol, making the option less trustworthy and was causing checks to fail due to the change on this PR.