Patch LibPoly for newer versions of Clang (#9266)
Fixes #9265. On newer versions of a Clang a variable, enabled_count,
in LibPoly leads to warnings, which aborts compilation. PR #9006 fixed
the issue for Emscripten. This commit removes all lines related to the
variable as a workaround for all compilers. In contrast to the
workaround in PR #9006, the workaround in this commit does not rely on
the name of the warning.