Track forced logic in the symbol manager (#9275)
This moves the tracking of the forced logic (--force-logic=LOGIC) from the Parser to the SymbolManager. It ensures that the TPTP parser handles forced logic on initialization instead of dynamically.

Further cleanup is done to the Parser class to remove unecessary methods.

This is a prerequisite for the new parser API. This ensures that forceLogic(...) does not need to be part of the new API for parsers.