Strengthen check for when care pairs can be discarded   (#9035)
This should lead to slightly smaller care graphs for some theories like sets where care pairs are computed for operators that have both shared and non-shared children. For example, (set.member x A) where x is shared and A is not.

This PR modifies a callback that checks whether two terms do not need to be cared about. Two terms (f t1 ... tn) and (f s1 ... sn) do not generate care pairs if we know some ti != si.

Previously, we were assuming that all arguments of a function to test where shared, and thus were insisting that they were trigger terms. However, we failed to recognize when non-shared terms were disequal. This adds this check. It furthermore catches the case of disequal constants (which are trivially disequal, but not explicitly disequal in the equality engine).

This issue was found by @mudathirmahgoub when comparing the standard implementation of compute care graphs with the array implementation.

Improves the size of the care graph for bags on the PR #8969. It should not impact performance significantly.

FYI @barrettcw