Fix check-unsat-cores and minimal-unsat-cores (#9286)
This fixes bugs related to checking unsat cores and minimal unsat cores.

Both features applied top-level substitutions to assertions-to-check, which includes definitions for define-fun and substitutions learned via non-clausal simplification. This may lead to incorrectly determining that an unsat core is unsatisfiable.

This meant that
(1) --minimal-unsat-cores could produce incorrect unsat cores
(2) --check-unsat-cores could fail to realize when an unsat core was incorrect.

The PR ensures that top-level substitutions are not applied, and assertions corresponding to definitions are treated via defineFunction instead of assertFormula on the subsolver.

This furthermore corrects an issue with unsat cores + --macro-quant where we were producing incorrect unsat cores but failing to realize it. It also disables unsat cores from 2 regressions where --check-unsat-cores now times out.

A minimized regression for minimal-unsat-cores is added.