Avoid lexing several theory-specific operators (#9301)
This is work towards the new parser. This PR eliminates (conditional) lexing for many theory-specific operators.

The remaining operators e.g. set.comprehension, lambda, fmf.card are harder to deal with since they impact the control flow of parsing.

Furthermore, it makes match unconditionally lexed, which is standard in SMT-LIB 2.6.