Move ANTLR specific file inclusion to base parser (#9290)
Most of the implementation of the smt2 / tptp parsers (parser::Smt2 and parser::Tptp) will be reused in the new parser.

This is work towards factoring the ANTLR-specific features out of these classes. In particular, it moves the include file features of these classes to the base class.

It also does some minor cleanup to the base parser class.