Decouple Assertions and AssertionsPipeline (#9268)
This decouples Assertions (which tracks the user assertions in the current context) and AssertionsPipeline (a set of assertions we pass to the preprocessor).

The SmtSolver now only tracks Assertions and does not own an AssertionsPipeline. As a result, methods that invoke preprocessing are migrated to SmtDriver.

The interface of SmtDriver is extended to explicitly manage assertions pipeline. Its getNextAssertions method is made more uniform so that it is called before all calls to check-sat, even the first one. The default implementation (SmtDriverSingleCall) tracks what should be added to the assertions pipeline based on a context-dependent index, which replaces the current buffering policy.

Further cleanup will be done to the interface of these classes in followup PRs.