Do not reset context when checking again with SmtDriver (#9313)
The current implementation of SmtDriver assumed that assertions should be cleared for the next check-sat, which was done by popping/pushing the contexts. This was wrong for deep restarts since this would also clear the user assertion list and top-level substitutions in the environment.

This meant that (1) models were incorrect when using deep restarts (when a variable was added to top-level substitutions and subsequently cleared), (2) check-models was ineffective when using deep restarts since the user assertion list was cleared.

This removes this behavior, meaning that deep-restarts generates correct models.

This required removing an assertion from the equality engine regarding when equality engines should be constructed.