Process lemma types. (#9321)
This PR fixes an issue where the types of lemmas are printed in SMT-LIB format instead of LFSC. This issue occurs now because we did not need to process lemma types before as the lemmas were part of the final proof and LFSC could figure out their types automatically. However, with --lfsc-flatten, the lemmas are checked, and their definition is erased, before the final proof. Therefore, LFSC needs to know their type to ensure that the final proof type checks.