Branch on option string names using hash map and switch-case (#9335)
We have about 450 options. There are functions that branch on their
string names. Previously, this was done with an if-else chain:

if (name == "OPTION_1") { return ... }
if (name == "OPTION_2") { return ... }
...
Now, we use a hashmap to enums, followed by a switch-case (jump table):

OptionEnum t = MAP.at(name);
switch (t) {
  case OPTION_1: return ...
  case OPTION_2: return ...
  ...
}
This change seems to reduce the runtime of get by about 5x: the unit test that I added previously ran in ~1.23s and now it runs in ~0.24s.