Always flush the command output (#9352)
This is work towards fixing #9257.

That issue still hangs during parsing in some cases due to issues in the ANTLR lexer (in particular, for sygus inputs).

In the new flex parser, we do not have issues with lexing. However, as suggested on that issue, we don't always flush the command output. This fixes the issue, thus, the flex parser works on that python script. The issue will be closed when the flex parser is merged and made default.