Add new utility methods to smt2 parser state (#9360)
This adds new methods to the smt2 parser state class in preparation for the flex parser.

In particular, since no theory-specific operators will be lexed, we will not have custom parsing rules for e.g. set.comprehension. Thus, we require a new notion of a "closure kind".

We also will parse indexed symbols that are indexed by symbols in a uniform way, requiring new variants for mkIndexedConstant and mkIndexedOp.