Cache lemmas in TheoryStrings (#9365)
This makes the inference manager of TheoryStrings cache the lemmas it sends, which is the standard behavior.

Caching lemmas revealed an issue in one of our lemma schemas (STRINGS_DEQ_NORM_EMP), which could send the same lemma more than once, which can lead to non-termination. Note this schema is very seldom used. This PR fixes that schema to ensure that the lemmas it sends have premises are explainable in the current context, implying that the lemma is not a duplicate. It makes minor fixes and refactoring to the utilities used by this schema.

This is +47-19 on SMT-LIB with a 30 second timeout.