Refactor extended function solver (#9374)
This makes several refactorings to the extended function solver. The main goal is to allow more flexibility for modifying our strategy for handling regular expressions.

In particular, this PR:

It makes the "reduced" set aware of the polarity of extended function predicates (e.g. contains and memberships). It improves our reductions for contains so that it is reduced context-independent + aware of polarity instead of marked inactive. This potentially avoids spurious computation of reductions.
Makes minor simplifications and corrections to shouldDoReduction.
This is in preparation for #9373.