Use `env` in shebang lines (#9391)
On some systems, such as the Linux distribution NixOS, Bash is not at /bin/bash. By using the env command in the shebang line (i.e., #!/usr/bin/env bash) it is possible to have scripts that are correctly executed independent of the location of Bash.

This changes the shebang lines in some scripts from /bin/bash to /usr/bin/env. In the case of the get-lfsc-checker script, the shebang line already uses env, but some scripts generated by get-lfsc-checker do not.

The competition scripts are not changed. I guess those are in the repository more for archival purposes. I also noticed that env is already used in most places. Hence, feel free to reject this pull request if /bin/bash was used on purpose in the scripts I touch here.