Refactor the regular expression solver (#9373)
This makes several refactorings to the regular expression solver. The main goal is to allow more flexibility for modifying our strategy for handling regular expressions.

In particular, this PR:

Makes use of the new methods in the regular expression solver introduced on previous PRs, which separated custom normal form evaluations out of RegexpSolver::checkUnfold, and moved them into a new method RegexpSolver::checkEvaluations, and made doUnfold its own method for unfolding a single membership.
It unifies RegexpSolver's ucache and ccache sets with the extended function solver's reduced and inactive sets respectively. This will avoid redundant computation for checking evaluation of regular expression memberships in the extended solver.
It refactors the strategy step calls to the subsolvers to take Theory::Effort levels for clarity.
+3-0 on SMT-LIB string logics, all 3 benchmarks close to time out.