Do not allow nested annotations in quantifier bodies (#9308)
According to the SMT-LIB standard, multiple patterns for a quantifier body should be given as multiple attributes annotated on the body. The ANTLR parser is currently more permissive by allowed nested annotations, the new parser will not support this.

In preparation for the new flex-based parser, this updates 2 regressions to follow the standard syntax. It also fixes the smt2 printer to not print nested annotations in the case there are multiple attributes on a quantified formula body.