Fix pattern printing for match terms (#9309)
Currently we print pattern terms in match cases as terms. This is not quite right, since patterns are constructor symbols applied to variables according to SMT-LIB syntax. This means that for match terms over parametric datatypes, we were printing terms due to applications of qualified identifiers. Moreover our ANTLR parser is permissive to accept this syntax, whereas the new parser will not be.

This corrects the issue.