[proofs] Adds option to optimize elaboration of macro resolution steps (#9414)
During post-processing we break each macro resolution steps into a series chain resolution + factoring steps. How many of these pairs we add is highfly influenced by how many "crowding literals" we have (i.e., literals that in the macro resolution have their duplicates implicitly deleted) and by how far down the chain they occur. This optimization is based on a heuristic to move these literals further down the chain when it is safe to do so.

In a challenge industrial benchmark this optimization leads to a 10x reduction in the expansion caused to the proof by the elaboration of macro resolution (2k added steps rather than 20k).