Simplify quantifier elimination preprocessing in Sygus solver (#9440)
This simplifies quantifier elimination preprocessing in Sygus. The implementation misused preRegisterTerm as a way to know when a conjecture existed.

It furthermore was overly complicated by the use of lemmas, instead of directly using the result of the preprocessing internally.

This is work towards new approaches for preregistration (which previously had broken this feature).

This also removes a thread local variable that guarded node visitor from being reentrant (introduced in 8a0c056#diff-57a747130dc7daf353aee11916d4b093b4b1ae6710c114d011326a2106bdfdbc). I don't see a good reason to do this. This is required since we spawn a subsolver during registration in the simplified approach introduced in this PR.