Add theory preregistrar (#9446)
This adds a module, theory preregistrar, which is responsible for calling TheoryEngine::preRegister for atoms allocated in the SAT solver. This sets up the atom as a target for theory propagation.

This PR adds two trivial policies, eager and lazy. The former is what is done currently: we preregister atoms immediately when they are allocated. The latter preregisters atoms just before they are asserted.

It also renames some methods preRegister -> notifySatVariable to be more accurate to the new possible behaviors.

Further PRs will introduce an algorithm where only relevant literals are preregistered.