Fix theory need check tracking (#9443)
This fixes a subtle bug in the theory proxy class which could lead to cases where FULL effort check was called before we had a complete SAT assignment.

The issue could arise when we are propagating at FULL effort with no remaining SAT decisions, and an asserted literal could trigger a new skolem definition to become active (when using decision=justification).

In this case, the skolem definition needs a SAT assignment before running a full effort check. However, we ran the FULL effort check anyways in this case. This was leading to debug-check-model failures where TheoryEngine would incorrectly think it was ready to answer "sat".

We had previously been avoiding this issue since we nearly always send (redundant) propagations back to the SAT solver for literals immediately as they are asserted. (However, #9432 may change this). This means the code was being lucky by always re-checking at FULL effort in cases where literals were asserted (which is a superset of cases where skolem definitions are activated).

This corrects the issue by changing FULL -> STANDARD check when this occurs and moreover ensuring we recheck.

This issue occurs relatively infrequently (the new block of code is triggered in only 12 regressions).