Eliminate capture avoiding substitution (#9390)
We have adopted the policy that shadowing should be handled in the rewriter, not during substitution.

This eliminates capture avoiding substitution method, which confuses this policy. It extends the rewriter for all closure operators to eliminate shadowing explicitly. This should make the implementation of substitution+rewriting more robust, as we no longer have to care about the source of substitutions.

Note that the quantifiers rewriter (for FORALL, EXISTS) already eliminates shadowing during one of its passes.