Avoid spurious equality engine trigger predicates (#9432)
Depends on #9443.

The purpose of this PR is to minimize redundant theory propagations.

This adds an option to check whether a literal is already asserted in the SAT solver when registering trigger predicates to the equality engine. In the case it is, it is not registered as a trigger predicate.