Add theory engine modules infrastructure (#9317)
This standardizes the interface for "theory engine modules" in the theory engine, which act like theory solvers (e.g. send lemmas) but are not associated with a theory.

The partition generator and the relevance manager / difficulty manager are current examples of theory engine modules, which are made to conform to the new interface.