Purify terms in branch and bound lemmas when necessary (#9507)
A recent commit bede046 introduced rewrites for eliminating arithmetic inequalities.

This leads to a subtle issue where branch-and-bound lemmas may be ineffective, since splitting on (>= (bv2nat x) N) may be rewritten to a predicate involving (bvugt x bvN), which belongs to BV, meaning that the integer value of (bv2nat x) is not refined.

This corrects the issue by ensuring that branch-and-bound lemmas are purified when the inequality would otherwise be eliminated. This is exclusively the case for bv2nat terms currently, although the test checks the rewritten form.

Fixes #9506.