Add new type checking algorithm (#9497)
Incorporates:

a new method preComputeType for inferring the type of a node at pretraversal. This is independent of gradual types and is an optimization. Followup PRs will add preComputeType for each type rule and auto-generate the type checking code for them; for now it is a placeholder.
the ability for type rules to return TypeNode::null and write to errOut instead of throwing exceptions. The behavior is preserved for now by always throwing the exception in NodeManager when a type rule does so. This is in preparation for algorithms (e.g. SyGuS + gradual types) that may speculatively construct non-well-sorted terms.