Move clone to the ProofNode class (#9654)
This is a relatively simple refactoring. Currently there is a clone method in the ProofNodeManager that does a deep copy of a ProofNode. That is, it generates a clone of a proof tree. However, this method uses nothing specific to the ProofNodeManager, but requires the user to have access to a ProofNodeManager.

This pull request moves the clone method to the ProofNode itself.