Distinguish ppStaticRewrite from normal ppRewrite (#9680)
This corrects a fundamental issue with how a rewrite involving bv2nat is handled, which can lead to incorrect models.

Recently, we added a rewrite that is based around transforming (>= (bv2nat x) c) to (bvugt x i) where i is the evaluation of (int2bv c), which has the advantage of eliminating the need for reasoning about bv2nat. This has been helpful in practice on user benchmarks.

However, the problem is that the (linear) arithmetic solver requires introducing inequalities of the above form. For example, it may introduce such an inequality to branch-and-bound on the value of (bv2nat x). Since the above rewrite applies, the arithmetic solver does not achieve its intention when sending a branch and bound lemma.

Notice that we already have this restriction on equalities (equalities cannot be preprocessed).

This PR makes it so that this rewrite is only applied to the input formulas, and not to lemmas. To do so, we distinguish two kinds of preprocessing rewrites: ppStaticRewrite (applies to input formulas only) and ppRewrite (applies to input formulas and lemmas).

It generalizes a preprocessing pass pp-rewrite-eq which previously applied to equalities only. This generalizes the functionality of this pass, and explicitly separates its callback into theory engine (ppStaticRewrite).

Followup PRs will move the preprocessing pass theory_rewrite_eq.h/cpp -> theory_static_pp_rewrite.h/cpp, and perform some simplifications.

Fixes #9660.
Fixes cvc5/cvc5-projects#623.