ci: Set PYTHON_* prior to calling find_package(PythonExtensions). (#9704)
The Interpreter only sets Python_EXECUTABLE, but not the development parts. The workaround in the top-level CMakeLists.txt however still set the INCLUDE_DIRS and LIBRARIES, which were empty at that stage. As a result these variables were empty but defined and not properly poplated in src/api/python/CMakeLists.txt when finding the Development components of Python.