Optimizations to parser (#9723)
This makes two optimizations to the parser:
(1) Cleans up the ParseOp class so that it stores less information. Allocation/deallocation of this class tends to be a bottleneck.
(2) Minimize the lookups in symbol table. Previously we checked for an error before looking up, now the error is computed when a lookup fails.

These lead to roughly 10% improvement on parsing time in initial tests.