Added CMake fixtures to have 'make test' work when the executable is not built. (#9750)
Fixes #9569.

We must ensure that the cvc5 binary is built before running any tests. In principle, this would just be a matter of adding a dependency on the 'test' target (defined by CTest). However, that built-in targets are not accessible via user code.

This seems to be a legacy cmake bug from +10 years ago (8774, 8438).

In order to have 'make test' work properly I use a CMake feature called FIXTURES to add 'test_setup' as the first test which builds the target 'build-tests'.