Improvements for handling str.to_code (#9773)
This PR changes when we apply injectivity (see C-Inj https://homepage.divms.uiowa.edu/~ajreynol/ijcar20a.pdf) for str.to_code.

In particular, it should only be necessary to apply injectivity for pairs of string equivalence classes that may have length 1 and that are disequal. Previously we applied it to add pairs of equivalence classes, apart from pairs that both contain constants.

This allows us to send significantly fewer lemmas in some cases and has a positive effect on a small number of benchmarks of interest.

It also changes so we don't do a reduction lemma (range constraint) for str.to_code applied to concrete strings, since we have a concrete integer equality in these cases.