[proofs] [sat] Cover missed case when tracking optimized clauses' proofs (#9679)
We must only track the clauses, whose levels are optimized, for which we are generating proofs. Since we do not generate proofs for redundant clauses we must avoid tracking in such cases. Previously we were only checking for when the clause already had a generator, but there are other tests in endResChain. This commit moves the place where we track to guarantee we only track after all the tests are done.

Fixes #9669