Fix registration of arithmetic proof checkers (#9808)
The registration of (non-linear) arithmetic proof checkers was wrong in a few subtle ways, (1) proof checkers would be registered prior to when TheoryEngine::initializeProofCheck, (2) proof checkers would not be re-registered if a (deep) restart was performed.

This also moves ownership of the "main" arithmetic proof checker from the linear solver to the main arithmetic solver. This is in preparation for exploring new backends for the linear solver.