Unify purification skolems and skolem functions (#9797)
This makings purification skolems a special instance of skolem functions. This means that all generated skolems are formalized in a uniform way.

As a result, purification skolems have mandatory names, e.g. SKOLEM_PURIFY_n for numerals n. This changes all calls to mkPurifySkolem to not provide a name/comment, as these were redundant and counterproductive (a name for a skolem would hide the fact that it was a purification skolem and could lead to name irregularities if a skolem came from two sources).

This is work towards potentially exporting/importing skolems through the API.