Reenable support for all versions of python  (#9844)
Fixes #9771.
Python support is no longer limited to 3.10. This is part of an ongoing change to upgrade the deprecated python bindings code we are currently calling.

As discussed offline, I temporarily disabled python bindings on CI for MacOS. There is a issue with brew install paths on all MacOS machines. This will be reenabled once the deprecated calls are replaced by updated ones.