Fix shadowed variable issue in beta reduction (#9880)
We removed support for general capture avoiding substitution in #9390, in favor of methods for eliminating shadowing explicitly.

This was not correct for cases where substitutions could be applied whose range has free variables, as capturing could still occur.

Due to our parser, this only cases in which a bad substitution could occur is in higher-order logic, where a handful of model unsoundness bugs were reported in a recent run of TPTP.

This ensures that variables in the arguments of beta reductions are eliminated from the bodies of the lambdas they substitute into.

Note this implies that we have to compute the free variables of terms that are arguments to beta reduction (e.g. arguments of define-fun). This is typically a cached single traversal over terms (to compute expr::hasBoundVar).