Use Boolean purify skolems instead of BOOLEAN_TERM_VARIABLE (#9853)
This removes the "BOOLEAN_TERM_VARIABLE" kind.

This kind is used to mark that a Boolean variable needs to be sent to the theory solvers, e.g. it occurs as the argument of a function.

Instead, we now distinguish such variables as any Boolean SKOLEM whose skolem function id is SkolemFunId::PURIFY.

This is work towards simplifying the interface for constructing skolems, which is required for importing/exporting them.

FYI @barrettcw

Fixes cvc5/cvc5-wishues#127.