Ensure standard ordering on (unrewritten) equalities from theory combination (#9855)
This is a minor optimization that ensures that we generate equalities in theory combination that match the ordering used by most theories. This avoids having symmetric versions of an equality both occurring as theory literals, which is not intuitive when debugging.

It should have very little impact on behavior (I observed almost no change in stats change on regressions with this change).