Ensure notifications for all options changes in set defaults (#9460)
This ensures that all changes to options in SetDefaults are notified in a uniform way.

The PR makes minor changes to defaults (e.g. the UF symmetry breaker is always auto-configured, but rarely is true, thus its default value is changed). It also changes the logic slightly concerning when we notify when options changes. We now notify whenever an option is changed (at verbosity=1), regardless of whether the change conflicts with a user option change.

This helps document the auto-configuration behavior of cvc5 for the user.