Sygus Grammar Update. (#9850)
This PR removes the preprocessing changes that were added to the grammar class. With this change, the grammar class is back to being just a simple datatype builder (with the exception of d_datatype kept for caching and user API purposes). The PR also reduces usage of SygusDatatype in preparation for removing it in the future.