[proofs] [alethe] Flatten HO chains (#9682)
When eliminating non-constant defined functions cvc5 will use HO_APPLY, in a curried manner, to represent applications. Since this kind is printed with "@" it clashes with what is expected in Alethe. This commit forces cvc5 to flatten the (note that total) application of the lambda representing the definition of the function. So for example ((lambda ((x1 Bool) (x2 Bool)) (and x1 x2)) p1 true) rather than (@ (@ (lambda ((x1 Bool) (x2 Bool)) (and x1 x2)) p1) true).